pyproject.toml
[project]
name = "langchain-rag1"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "chromadb>=1.5.8",
    "ipykernel>=7.2.0",
    "langchain>=1.2.15",
    "langchain-chroma>=1.1.0",
    "langchain-openai>=1.1.14",
    "langchain-text-splitters>=1.1.2",
    "python-dotenv>=1.2.2",
]

[dependency-groups]
dev = [
    "pre-commit>=4.5.1",
    "rust-just>=1.49.0",
]
